% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coga.R
\docType{package}
\name{coga}
\alias{coga}
\alias{coga-package}
\title{coga: Convolution of independent Gamma Distributions}
\description{
Evaluation for density and distribution function of convolution of gamma
distributions in R. Two related exact methods and one approximate method are
implemented with efficient algorithm and C++ code. A quick guide for choosing
correct method and usage of this package is given in package vignette.
}
\section{coga functions}{

dcoga, pcoga, rcoga, dcoga2dim, pcoga2dim, pcoga2dim_diff_shape, dcoga_approx, and
pcoga_approx.
}

\author{
\itemize{
  \item *maintainer, author* Chaoran Hu \email{huchaoran.stat@gmail.com}
  \item *advisor* Jun Yan \email{jun.yan@uconn.edu}
  \item *advisor* Vladimir Pozdnyakov \email{vladimir.pozdnyakov@uconn.edu}
}
}
