% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcoga}
\alias{dcoga}
\alias{dcovgam}
\title{Density function of Convolution of Gamma Distributions.}
\usage{
dcoga(x, shape, rate)
}
\arguments{
\item{x}{Quantiles.}

\item{shape}{Numerical vector of shape parameters of every gamma distributions,
all shape parameters >= 0, at least one shape parameter > 0.}

\item{rate}{Numerical vector of rate parameters of every gamma distributions.}
}
\value{
The density at x.
}
\description{
The distribution of the convolution of independent Gamma random variables with
different parameters is \eqn{Y = X_{1} + ... + X_{n}}, where \eqn{X_{i}, i = 1, ..., n},
are independent Gamma distributions with parameters shapes and rates.
The density function can be calculated, according to the formulas
from Moschopoulos, Peter G. (1985).
}
\examples{
y <- rcoga(100000, c(3,4,5), c(2,3,4))
grid <- seq(0, 15, length.out=100)
pdf <- dcoga(grid, shape=c(3,4,5), rate=c(2,3,4))

plot(density(y), col="blue")
lines(grid, pdf, col="red")

}
\references{
Moschopoulos, Peter G. "The distribution of the sum of independent
gamma random variables." Annals of the Institute of
Statistical Mathematics 37.1 (1985): 541-544.
}
\author{
Chaoran Hu
}
