% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprcoga.R
\name{rcoga}
\alias{rcoga}
\alias{rcovgam}
\title{Random Generation for Convolution of Gamma Distributions.}
\usage{
rcoga(n, shape, rate)
}
\arguments{
\item{n}{The number of observations, should be larger than 0.}

\item{shape}{Numerical vector of shape parameters of every gamma distributions,
all shape parameters >= 0, at least one shape parameter > 0.}

\item{rate}{Numerical vector of rate parameters of every gamma distributions.}
}
\value{
The random samples from given convolution of gamma distributions.
}
\description{
Random generation function for convolution of gamma distributions. Each gamma distribution
here can have different shape parameters and rate parameters.
}
\examples{
rcoga(10, c(2,0,0), c(2,3,1))

}
\author{
Chaoran Hu
}
